+++ src/main/java/de/siegmar/logbackgelf/GelfTcpTlsAppender.java
@@ -23,6 +23,8 @@ import java.io.ByteArrayInputStream;
 import java.nio.charset.StandardCharsets;
 import java.security.GeneralSecurityException;
 import java.security.KeyManagementException;
+import java.security.KeyStore;
+import java.security.KeyStoreException;
 import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
 import java.security.cert.CertificateException;
@@ -34,23 +36,24 @@ import java.util.List;
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.SSLSocketFactory;
 import javax.net.ssl.TrustManager;
+import javax.net.ssl.TrustManagerFactory;
 import javax.net.ssl.X509TrustManager;
 
 public class GelfTcpTlsAppender extends GelfTcpAppender {
 
     /**
-     * If {@code true}, trust all TLS certificates (even self signed certificates).
+     * If {@code true}, skip the TLS certificate validation.
      */
-    private boolean trustAllCertificates;
+    private boolean insecure;
 
     private List<X509Certificate> trustedServerCertificates = new ArrayList<>();
 
-    public boolean isTrustAllCertificates() {
-        return trustAllCertificates;
+    public boolean isInsecure() {
+        return insecure;
     }
 
-    public void setTrustAllCertificates(final boolean trustAllCertificates) {
-        this.trustAllCertificates = trustAllCertificates;
+    public void setInsecure(final boolean insecure) {
+        this.insecure = insecure;
     }
 
     public void addTrustedServerCertificate(final String trustedServerCertificate)
@@ -67,32 +70,42 @@ public class GelfTcpTlsAppender extends GelfTcpAppender {
 
     @Override
     protected SSLSocketFactory initSocketFactory() {
-        if (!trustedServerCertificates.isEmpty()) {
-            if (trustAllCertificates) {
-                throw new IllegalStateException("TrustAllCertificates is not possible when "
-                    + "configuring server and/or CA certificates explicitly");
+        return configureSocket();
+    }
+
+    private SSLSocketFactory configureSocket() {
+        final TrustManager trustManager;
+
+        try {
+            if (insecure) {
+                addWarn("Enabled insecure mode (skip TLS certificate validation)"
+                    + " - don't use this in production!");
+                trustManager = new NoopX509TrustManager();
+            } else {
+                trustManager = new CustomX509TrustManager(defaultTrustManager(), getGraylogHost(),
+                    trustedServerCertificates);
             }
 
-            try {
-                final EasyX509TrustManager trustManager = new EasyX509TrustManager();
-                trustManager.setTrustedServerCertificates(trustedServerCertificates);
+            return configureSslFactory(trustManager);
+        } catch (final GeneralSecurityException e) {
+            throw new IllegalStateException(e);
+        }
+    }
 
-                return configureSslFactory(trustManager);
-            } catch (final GeneralSecurityException e) {
-                throw new IllegalStateException(e);
+    private static X509TrustManager defaultTrustManager()
+        throws NoSuchAlgorithmException, KeyStoreException {
+
+        final TrustManagerFactory trustManagerFactory =
+            TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm());
+        trustManagerFactory.init((KeyStore) null);
+
+        for (final TrustManager trustManager : trustManagerFactory.getTrustManagers()) {
+            if (trustManager instanceof X509TrustManager) {
+                return (X509TrustManager) trustManager;
             }
         }
 
-        if (trustAllCertificates) {
-            addWarn("Enable trustAllCertificates - don't use this in production!");
-            try {
-                return configureSslFactory(buildNoopTrustManager());
-            } catch (final NoSuchAlgorithmException | KeyManagementException e) {
-                throw new IllegalStateException(e);
-            }
-        }
-
-        return (SSLSocketFactory) SSLSocketFactory.getDefault();
+        return null;
     }
 
     private SSLSocketFactory configureSslFactory(final TrustManager trustManager)
@@ -101,20 +114,6 @@ public class GelfTcpTlsAppender extends GelfTcpAppender {
         final SSLContext context = SSLContext.getInstance("TLS");
         context.init(null, new TrustManager[]{trustManager}, new SecureRandom());
         return context.getSocketFactory();
-    }
-
-    private static TrustManager buildNoopTrustManager() {
-        return new X509TrustManager() {
-            public void checkClientTrusted(final X509Certificate[] chain, final String authType) {
-            }
-
-            public void checkServerTrusted(final X509Certificate[] chain, final String authType) {
-            }
-
-            public X509Certificate[] getAcceptedIssuers() {
-                return new X509Certificate[0];
-            }
-        };
     }
 
 }